requirements.txt
# Core ML/DL libraries
tensorflow==2.20.0
keras==3.12.0
numpy==2.3.4

# CLIP and text encoding (no PyTorch dependency)
sentence-transformers>=2.2.0
transformers>=4.30.0

# Image processing
opencv-python>=4.8.0
pillow==12.0.0

# Utilities
tqdm==4.67.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2

# TensorFlow dependencies
absl-py==2.3.1
astunparse==1.6.3
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
libclang==18.1.1
ml_dtypes==0.5.3
namex==0.1.0
opt_einsum==3.4.0
optree==0.17.0
protobuf==6.33.0
termcolor==3.2.0
wrapt==2.0.1

# Other dependencies
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
idna==3.11
kagglehub==0.3.13
kiwisolver==1.4.9
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
packaging==25.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
